+++ docs/app/views/components/examples/tooltips/align_end_component.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module Examples
+  module Tooltips
+    class AlignEndComponent < Base
+      def title
+        "Align end"
+      end
+
+      def example
+        Tooltip :open, :top, :end, tip: "hello" do
+          Button do
+            "End"
+          end
+        end
+      end
+    end
+  end
+end