+++ docs/app/views/examples/auras/show_view.rb
@@ -1,20 +1,19 @@
 # frozen_string_literal: true
 
 module Examples
-  module Ranges
+  module Auras
     class ShowView < BaseView
       def view_template
         title do
-          "Range"
+          "Aura"
         end
 
         render_examples [
           BasicComponent,
-          WithColorsComponent,
+          AroundButtonComponent,
+          RainbowComponent,
+          StylesComponent,
           SizesComponent,
-          WithStepsComponent,
-          WithLabelComponent,
-          VerticalComponent,
         ]
       end
     end