+++ docs/app/views/examples/otps/show_view.rb
@@ -1,20 +1,19 @@
 # frozen_string_literal: true
 
 module Examples
-  module Ranges
+  module Otps
     class ShowView < BaseView
       def view_template
         title do
-          "Range"
+          "OTP"
         end
 
         render_examples [
           BasicComponent,
+          JoinedComponent,
+          SixDigitsComponent,
           WithColorsComponent,
           SizesComponent,
-          WithStepsComponent,
-          WithLabelComponent,
-          VerticalComponent,
         ]
       end
     end